bug_fixes

+++ src/color.c
@@ -9,7 +9,7 @@ int clamp(int a)
 	return (a);
 }
 
-int		color_parse(t_color col, float c)
+int		color_parse(t_color col, double c)
 {
 	int r;
 	int g;
@@ -22,21 +22,3 @@ int		color_parse(t_color col, float c)
 	color = (r << 16 | g << 8 | b);
 	return (color);
 }
-
-// void    sphere_mapping_texture(t_additional *s, void *t)
-// {
-//     t_rtv                *rtv;
-//     t_mapping_texture    *tx;
-//     t_sphere            *sp;
-//     t_vect                temp;
-
-//     rtv = (t_rtv *)t;
-//     sp = (t_sphere*)rtv->objects[s->figure].object;
-//     tx = (t_mapping_texture *)rtv->objects[s->figure].texture.tx_struct;
-//     temp = vector_substract(&s->point, &sp->pos);
-//     temp = normalize_vector(&temp);
-//     VAR_INT(i, 0);
-//     i = (int)((0.5 + (atan2(temp.z, temp.x) / (2.0 * M_PI))) * 
-// 	tx->w) + tx->w * (int)((0.5 - (asin(temp.y) / M_PI)) * tx->h);
-//     s->mat.cl = parse_color(tx->arr[i]);
-// }
